# Moving iteration over 2nd axis under function submitted to a thread pool for 3d EDT speeds up execution time to up to 3x