@william-silversmith 2024-02-18
Thanks so much for this PR. Sorry the initial notifications got lost in my inbox. Thank you for your persistence! I tried this on some real data and got some nice results, including the speedup for 1 thread that you saw as well.

```
# old results
12.29  sec. t=1
 6.51  sec. t=2
 5.17  sec. t=3
 4.73  sec. t=4
 4.41  sec. t=5
 4.02  sec. t=6
 3.78  sec. t=7
 3.68  sec. t=0
```

```
# new results
11.98  sec. t=1
 6.04  sec. t=2
 4.78  sec. t=3
 4.29  sec. t=4
 3.72  sec. t=5
 3.32  sec. t=6
 3.02  sec. t=7
 2.94  sec. t=0
```

Your rationale makes sense to me as well. I knew the units of work were small, but for some reason it didn't occur to me to try this.

Thank you for a great contribution!